refactor dense_decomposition and bigfish.detection
