+++ bigfish/detection/__init__.py
@@ -12,13 +12,14 @@ from .spot_detection import local_maximum_detection
 from .spot_detection import spots_thresholding
 from .spot_detection import automated_threshold_setting
 
-from .dense_decomposition import decompose_cluster
-from .dense_decomposition import fit_subpixel
-from .dense_decomposition import build_reference_spot
-from .dense_decomposition import modelize_spot
-from .dense_decomposition import precompute_erf
-from .dense_decomposition import get_clustered_region
-from .dense_decomposition import fit_gaussian_mixture
+from .dense_decomposition import decompose_dense
+from .dense_decomposition import get_dense_region
+from .dense_decomposition import simulate_gaussian_mixture
+
+from .spot_modeling import fit_subpixel
+from .spot_modeling import build_reference_spot
+from .spot_modeling import modelize_spot
+from .spot_modeling import precompute_erf
 
 from .cluster_detection import detect_clusters
 
@@ -30,15 +31,17 @@ _spots = [
     "automated_threshold_setting"]
 
 _dense = [
-    "decompose_cluster",
+    "decompose_dense",
+    "get_dense_region",
+    "simulate_gaussian_mixture"]
+
+_model = [
     "fit_subpixel",
     "build_reference_spot",
     "modelize_spot",
-    "precompute_erf",
-    "get_clustered_region",
-    "fit_gaussian_mixture"]
+    "precompute_erf"]
 
 _clusters = [
     "detect_clusters"]
 
-__all__ = _spots + _dense + _clusters
+__all__ = _spots + _dense + _model + _clusters
