+++ bigfish/detection/tests/test_dense_decomposition.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+# Author: Arthur Imbert <arthur.imbert.pro@gmail.com>
+# License: BSD 3 clause
+
+"""
+Unitary tests for bigfish.detection.dense_decomposition module.
+"""
+
+# TODO test bigfish.detection.decompose_dense
+# TODO test bigfish.detection.get_dense_region
+# TODO test bigfish.detection.simulate_gaussian_mixture
