+++ bigfish/detection/tests/test_spot_detection.py
@@ -3,9 +3,10 @@
 # License: BSD 3 clause
 
 """
-Unitary tests for bigfish.stack.spot_detection module.
+Unitary tests for bigfish.detection.spot_detection module.
 """
 
-# TODO test bigfish.detection.spot_detector
+# TODO test bigfish.detection.detect_spots
 # TODO test bigfish.detection.local_maximum_detection
 # TODO test bigfish.detection.spots_thresholding
+# TODO test bigfish.detection.automated_threshold_setting
