+++ bigfish/detection/tests/test_spot_modeling.py
@@ -3,12 +3,10 @@
 # License: BSD 3 clause
 
 """
-Unitary tests for bigfish.stack.cluster_decomposition module.
+Unitary tests for bigfish.detection.spot_modeling module.
 """
 
-# TODO test bigfish.detection.decompose_cluster
+# TODO test bigfish.detection.fit_subpixel
 # TODO test bigfish.detection.build_reference_spot
 # TODO test bigfish.detection.modelize_spot
 # TODO test bigfish.detection.precompute_erf
-# TODO test bigfish.detection.get_clustered_region
-# TODO test bigfish.detection.fit_gaussian_mixture
